Fixes issue with query reset
We were hitting a validation error that we had used the query without
resetting it. I was resetting it on the host, but it's possible for a command
using it to be queued up and run on the device after the reset.

I'm going to see if I can do a followup where I also replace the pool with
a ring buffer or something.