Moves away from draw indirect
It's actually significantly slower (12x!) for our use case. We used to
be using it more optimally, but now we're only actually issuing one draw
per batch anyway.

That doesn't explain why it's so slow. Could be an issue on our end or could
be poor driver support (testing on Linux atm), either way it's much *simpler*
to draw directly so I'll take a speedup this big alongside simpler code any
day.