Removes backend interface struct, see long notes
This was a nice idea. It made clear what functions and types backends
had to provide, making it easier to implement new backends and, and to
evolve the interface clearly understanding when changes are breaking.

However, it was resulting in a lot of really annoying dependency loops.
This cost was outweighting the potential future benefits, so I removed it.
The new rule is that anything public, you have to assume the frontend may
be allowed to depend on.

If we actually do want to expose things from the backend that aren't relevant
to the frontend some day, we can revisit this or just come up with a naming
convention.