Adds support for push constants.
IMO it's generally better to just make a large buffer with all args and
index on the instance ID. That way you have overall fewer draw calls.
However, this isn't possible with compute since there's no instance ID,
and we need a way to pass args or indices into our arg arrays to post
processing shaders and such.

(You could try to re purpose the z component of the invocation ID or
something, but that's a bit hacky since now it can't be used for
its intended purpose as loose as it is, and there's no equivalent DX12
for that anyway since AFAICT you can't set the base invocation ID in
DX12 like you can in Vulkan.)