Returns image indices instead of image views from acquire next image
We want the end user to have the option to get the full array up front,
as they may want to e.g. bind it to a descriptor as an array and index it
frame to frame in a compute shader to decide what to write to.

As part of this, the swapchain image transitions are no longer handled
automatically. This didn't really make sense because it assumed you were
writing the image from the graphics pipeline as a color attachment, when
it's not uncommon to actually write the final image with a compute pipeline
for post processing purposes instead.

This change is a work in progress insofar as it currently exposes the arrays
and extent via a getter, instead of just exposing the array, and it also
doesn't currently provide a way to know when the swapchain was recreated which
is needed to know when to update a descriptor with an array of the image views.