Removes default value for immutable samplers, removes combined image samplers
Immutable samplers appear to be significantly faster on integrated GPUs, you
should have to explicitly acknowledge that you're not using them if you for
some reason need this to be dynamic.

Combined image samplers don't appear to be very useful, I think they likely only
exist in the API for back compat.