First pass at using vendor specific input latency extensions
AMD's seems to be broken. If I don't set an fps limit, which is allegedly allowed,
it either fails to sleep or sleeps before rendering instead of before polling input.
I'm going to implement the equivalent Nvidia extension to see if that gives me any
insight into what's going wrong, it may just be broken or I may be missing something.