Don't default enable minimal validation or release safe builds
In theory this makes sense, it's a smaller perf impact and makes the
build safer, but it doesn't matter since you're likely redistriubting
release safe builds to setups that don't have validation layers installed.

If you want this behavior, you can still enable it by just not using the
defaults of course.